requirement.txt
absl-py>=0.13.0
addict>=2.4.0
aiohttp 
alabaster>=0.7.12
alembic>=1.7.5
allennlp>=2.6.0
anytree>=2.8.0
appdirs>=1.4.4
argon2-cffi 
async-generator>=1.10
async-timeout>=3.0.1
attrs 
autopage>=0.4.0
Babel>=2.9.1
backcall 
backports.csv>=1.0.7
beautifulsoup4>=4.9.3
bleach 
blis>=0.4.1
block.bootstrap.pytorch>=0.1.6
bootstrap.pytorch>=0.0.13
boto3 
botocore 
bpemb>=0.3.3
brotlipy>=0.7.0
cachetools 
catalogue>=2.0.4
certifi>=2021.5.30
cffi 
chardet 
charset-normalizer 
checklist>=0.0.11
cheroot>=8.5.2
CherryPy>=18.6.1
cityscapesScripts>=2.2.0
click 
cliff>=3.10.0
cloudpickle 
cmaes>=0.8.2
cmd2>=2.3.3
colorama 
coloredlogs>=15.0.1
colorlog>=6.6.0
configparser>=5.0.2
contextvars>=2.4
coverage 
cryptography 
cupy>=9.0.0
cxxfilt>=0.2.2
cycler>=0.10.0
cymem>=2.0.5
Cython 
cytoolz>=0.11.0
dask 
dataclasses 
de-core-news-sm 
decorator 
defusedxml 
Deprecated 
dill>=0.3.4
docker-pycreds>=0.4.0
docutils>=0.17.1
easydict>=1.9
einops>=0.3.2
en-core-web-md 
en-core-web-sm 
entrypoints>=0.3
fastrlock>=0.6
feedparser>=6.0.8
filelock>=3.0.12
flake8>=3.9.2
future>=0.18.2
gensim 
gitdb>=4.0.7
GitPython>=3.1.18
google-api-core 
google-auth 
google-auth-oauthlib>=0.4.4
google-cloud-core 
google-cloud-storage 
google-crc32c 
google-resumable-media 
googleapis-common-protos 
graphql-core>=3.1.5
greenlet>=1.1.2
grpcio>=1.38.0
h5py>=2.10.0
huggingface-hub>=0.0.12
humanfriendly>=9.2
idna 
idna-ssl 
imagecodecs 
imageio 
imagesize>=1.2.0
immutables>=0.15
importlib-metadata 
importlib-resources>=5.2.2
ipdb>=0.13.6
ipykernel 
ipython 
ipython-genutils 
ipywidgets 
iso-639>=0.4.5
jaraco.classes>=3.2.1
jaraco.collections>=3.3.0
jaraco.functools>=3.3.0
jaraco.text>=3.5.1
jedi>=0.18.0
Jinja2 
jmespath 
joblib 
jsonschema 
jupyter>=1.0.0
jupyter-client 
jupyter-console 
jupyter-core 
jupyterlab-pygments 
jupyterlab-widgets 
kiwisolver 
kornia>=0.5.6
llvmlite>=0.36.0
lmdb>=1.2.1
lxml>=4.6.3
Mako>=1.1.6
Markdown>=3.3.4
MarkupSafe 
matplotlib 
mccabe>=0.6.1
mistune>=0.8.4
mkl-fft>=1.3.0
mkl-random>=1.1.1
mkl-service>=2.3.0
mmcls>=0.10.0
mmcv>=1.3.1
multidict 
munch>=2.5.0
murmurhash>=1.0.5
nbclient 
nbconvert 
nbformat 
nest-asyncio 
networkx 
ninja>=1.10.2.2
nltk 
notebook 
numba 
numpy>=1.19.5
oauthlib>=3.1.1
olefile>=0.46
opencv-python>=4.5.1.48
optuna>=2.10.0
overrides>=3.1.0
packaging 
pandas 
pandocfilters 
parso 
pathlib>=1.0.1
pathtools>=0.1.2
pathy>=0.6.0
patternfork-nosql>=3.6
pbr>=5.8.0
pdfminer.six>=20201018
pickleshare 
Pillow 
plac>=1.1.3
plotly>=4.14.3
portend>=2.7.1
prefetch-generator>=1.0.1
preshed>=3.0.5
pretrainedmodels>=0.7.4
prettytable>=2.4.0
prometheus-client 
promise>=2.3
prompt-toolkit 
protobuf>=3.17.2
psutil>=5.8.0
pyasn1 
pyasn1-modules>=0.2.8
pycocotools>=2.0.2
pycodestyle>=2.7.0
pycparser 
pydantic>=1.8.2
pyflakes>=2.3.1
Pygments 
pyOpenSSL 
pyparsing 
pyperclip>=1.8.2
pyquaternion>=0.9.9
pyreadline>=2.1
pyrsistent 
PySocks 
pytest>=6.2.4
python-dateutil 
python-docx>=0.8.11
pytorch-transformers>=1.2.0
pytz 
PyWavelets 
pywin32>=228
pywinpty>=0.5.7
PyYAML>=5.4.1
pyzmq 
qtconsole 
QtPy 
regex 
requests 
requests-oauthlib>=1.3.0
retrying>=1.3.3
rsa 
s3transfer 
sacremoses>=0.0.45
scikit-image>=0.17.2
scikit-learn>=0.24.2
scipy>=1.5.4
seaborn>=0.11.1
segtok>=1.5.10
Send2Trash 
sentencepiece>=0.1.96
sentry-sdk>=1.3.1
sgmllib3k>=1.0.0
shortuuid>=1.0.1
six 
skipthoughts>=0.0.1
sklearn>=0.0
smart-open 
smmap>=4.0.0
snowballstemmer>=2.1.0
sortedcontainers>=2.4.0
soupsieve>=2.2.1
spacy>=2.2.2
spacy-legacy>=3.0.8
Sphinx>=4.0.2
sphinxcontrib-applehelp>=1.0.2
sphinxcontrib-devhelp>=1.0.2
sphinxcontrib-htmlhelp>=2.0.0
sphinxcontrib-jsmath>=1.0.1
sphinxcontrib-qthelp>=1.0.3
sphinxcontrib-serializinghtml>=1.1.5
SQLAlchemy>=1.4.28
srsly>=1.0.5
stanfordcorenlp>=3.9.1.1
stevedore>=3.5.0
tabulate>=0.8.9
tempora>=4.1.1
tensorboard>=2.5.0
tensorboard-data-server>=0.6.1
tensorboard-plugin-wit>=1.8.0
tensorboardX>=2.1
termcolor>=1.1.0
terminado>=0.9.4
testpath 
Theano>=1.0.5
thinc>=7.3.1
thop>=0.0.31.post2005241907
threadpoolctl>=2.2.0
tifffile 
tokenizers>=0.10.3
toolz
tornado 
tqdm 
traitlets 
transformers>=4.8.2
typer>=0.3.2
typing>=3.7.4.3
typing-extensions 
urllib3>=1.26.7
wandb>=0.11.1
wasabi>=0.8.2
wcwidth 
webencodings>=0.5.1
Werkzeug>=2.0.1
widgetsnbextension>=3.5.1
win-inet-pton 
wincertstore>=0.2
wrapt>=1.12.1
xdg>=5.1.0
yacs>=0.1.8
yapf>=0.31.0
yarl 
zc.lockfile>=2.0
zipp 
